REFACTOR: Move TypeBuilder (cardinality functionality) to parse module. Cleanup code and remove duplicated code sections.
